nfdiv-1183: move constants to CommonContent class
