+++ src/main/java/uk/gov/hmcts/divorce/notification/CommonContent.java
@@ -13,25 +13,42 @@ import java.util.Map;
 
 import static java.lang.String.join;
 import static uk.gov.hmcts.divorce.notification.FormatUtil.formatId;
-import static uk.gov.hmcts.divorce.notification.NotificationConstants.APPLICANT_NAME;
-import static uk.gov.hmcts.divorce.notification.NotificationConstants.APPLICATION_REFERENCE;
-import static uk.gov.hmcts.divorce.notification.NotificationConstants.COURT_EMAIL;
-import static uk.gov.hmcts.divorce.notification.NotificationConstants.DISSOLUTION_COURT_EMAIL;
-import static uk.gov.hmcts.divorce.notification.NotificationConstants.DIVORCE_COURT_EMAIL;
-import static uk.gov.hmcts.divorce.notification.NotificationConstants.FIRST_NAME;
-import static uk.gov.hmcts.divorce.notification.NotificationConstants.IS_DISSOLUTION;
-import static uk.gov.hmcts.divorce.notification.NotificationConstants.IS_DIVORCE;
-import static uk.gov.hmcts.divorce.notification.NotificationConstants.LAST_NAME;
-import static uk.gov.hmcts.divorce.notification.NotificationConstants.NO;
-import static uk.gov.hmcts.divorce.notification.NotificationConstants.PARTNER;
-import static uk.gov.hmcts.divorce.notification.NotificationConstants.RESPONDENT_NAME;
-import static uk.gov.hmcts.divorce.notification.NotificationConstants.SIGN_IN_DISSOLUTION_URL;
-import static uk.gov.hmcts.divorce.notification.NotificationConstants.SIGN_IN_DIVORCE_URL;
-import static uk.gov.hmcts.divorce.notification.NotificationConstants.SIGN_IN_URL_NOTIFY_KEY;
-import static uk.gov.hmcts.divorce.notification.NotificationConstants.YES;
 
 @Component
 public class CommonContent {
+
+    public static final String PARTNER = "partner";
+    public static final String FIRST_NAME = "first name";
+    public static final String LAST_NAME = "last name";
+
+    public static final String IS_DIVORCE = "isDivorce";
+    public static final String IS_DISSOLUTION = "isDissolution";
+
+    public static final String IS_REMINDER = "isReminder";
+    public static final String ACTION_REQUIRED = "actionRequired";
+    public static final String YES = "yes";
+    public static final String NO = "no";
+
+    public static final String CREATE_ACCOUNT_LINK = "create account link";
+    public static final String SIGN_IN_URL_NOTIFY_KEY = "signin url";
+    public static final String SIGN_IN_DIVORCE_URL = "signInDivorceUrl";
+    public static final String SIGN_IN_DISSOLUTION_URL = "signInDissolutionUrl";
+    public static final String DIVORCE_COURT_EMAIL = "divorceCourtEmail";
+    public static final String DISSOLUTION_COURT_EMAIL = "dissolutionCourtEmail";
+
+    public static final String COURT_EMAIL = "court email";
+
+    public static final String SUBMISSION_RESPONSE_DATE = "date of response";
+    public static final String APPLICATION_REFERENCE = "reference number";
+
+    public static final String ACCESS_CODE = "access code";
+
+    public static final String APPLICANT_NAME = "applicant name";
+    public static final String RESPONDENT_NAME = "respondent name";
+    public static final String SOLICITOR_NAME = "solicitor name";
+
+    public static final String CCD_REFERENCE = "CCD reference";
+    public static final String REVIEW_DEADLINE_DATE = "review deadline date";
 
     @Autowired
     private EmailTemplatesConfig config;
